+++ settings.example.py
@@ -3,10 +3,13 @@
 # General settings.
 main = {
 
-    # Measurement interval in seconds.
-    # TODO: Please note this is not the _real thing_ yet at it will just use
+    # Measurement intervals in seconds.
+    # Todo: Please note this is not the _real thing_ yet at it will just use
     #       this value to apply to ``time.sleep()`` after each duty cycle.
-    'interval': 15.0,
+    'interval': {
+        'field': 60.0,
+        'maintenance': 5.0,
+    },
 
     # Whether to use deep sleep between measurement cycles.
     'deepsleep': True,